Renamed class, added a BC extention
